requirements.txt
nltk~=3.9.1
python-docx~=1.1.2
scikit-learn~=1.6.0